chore(userspace): rename to common/types.h
