+++ userspace/libscap/scap.h
@@ -62,7 +62,7 @@ struct iovec;
 #include <time.h>
 #endif
 #include "uthash.h"
-#include "../common/sysdig_types.h"
+#include "../common/types.h"
 #include "../../driver/ppm_events_public.h"
 #ifdef _WIN32
 #include <time.h>
