+++ userspace/libscap/syscall_info_table.c
@@ -17,7 +17,7 @@ limitations under the License.
 
 */
 
-#include "../common/sysdig_types.h"
+#include "../common/types.h"
 #include "../../driver/ppm_events_public.h"
 #include "scap.h"
